Resolve compilation warning in main.cpp

+++ pa1/main.cpp
@@ -125,7 +125,7 @@ int main(int argc, char **argv) {
     Scene world{loader.LoadedMeshes[0]};
 
     // Create a renderer on scene
-    Zbuf zbuf{world, height, width};
+    Zbuf zbuf{world, static_cast<size_t>(height), static_cast<size_t>(width)};
 
     // Set fragment shader
     zbuf.set_shader(selected_fragment_shader);
